Modify issues about #48. 1. Rename GithubIssues component. 2. Using ES6 templete literals. 3. Replace tag to <p> from <br>. 4. Move directory about GithubIusse component and Issues container. 5. Add / in link path. 6. Clean up import order.
