+++ src/components/Github/GithubIssue/GithubIssue.js
@@ -6,9 +6,9 @@ class GithubIssues extends React.Component {
   render() {
     return (
       <div>
-        <Panel header={"Number : " + this.props.number} bsStyle="primary">
-          Title : {this.props.title} <br />
-          Body : {this.props.body}
+        <Panel header={`Number : ${this.props.number}`} bsStyle="primary">
+          <p> Title : {this.props.title} </p>
+          <p> Body : {this.props.body} </p>
         </Panel>
       </div>
     )
