+++ src/components/index.scss
@@ -6,4 +6,4 @@
 @import './Photos/PhotoDetail';
 @import './Photos/PhotoItem';
 @import './Albums/AlbumItem';
-@import './GithubIssues/GithubIssues';
+@import 'Github/GithubIssue/GithubIssue';
