+++ src/containers/About/About.js
@@ -3,12 +3,11 @@ import { Link } from 'react-router';
 import { Button } from 'react-bootstrap';
 
 class About extends React.Component {
-
   render() {
     return (
       <div>
         <h1 className="text-center"> ABOUT PAGE </h1>
-        <Link to="about/issues">
+        <Link to="/about/issues">
           <Button bsStyle="link"> ISSUES </Button>
         </Link>
         {this.props.children}
