+++ src/containers/index.js
@@ -1,7 +1,7 @@
 export App from './App/App';
 export Navbar from './Header/Header';
 export About from './About/About';
-export Issues from './About/Issues/Issues';
+export Issues from './About/Issues';
 export Post from './Posts/Post';
 export Posts from './Posts/Posts';
 export PostNew from './Posts/PostNew';
