+++ src/containers/index.scss
@@ -1,7 +1,7 @@
 @import './App/App';
 @import './Header/Header';
 @import './About/About';
-@import './About/Issues/Issues';
+@import 'About/Issues';
 @import './Posts/Post';
 @import './Posts/Posts';
 @import './Posts/PostNew';
